Separate starters for Spring Boot 1 and 2
