--- spring-boot/boot1-autoconfigure/pom.xml
@@ -1,98 +1,0 @@
-<?xml version="1.0"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.togglz</groupId>
-    <artifactId>togglz-spring-boot</artifactId>
-    <version>2.6.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>togglz-spring-boot1-autoconfigure</artifactId>
-  <name>Togglz - Spring Boot 1.5 Autoconfigure</name>
-  <description>Togglz - Spring Boot 1.5 Autoconfigure</description>
-
-  <properties>
-    <spring-boot.version>1.5.12.RELEASE</spring-boot.version>
-  </properties>
-
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <!-- Import dependency management from Spring Boot -->
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-dependencies</artifactId>
-        <version>${spring-boot.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-
-  <developers>
-    <developer>
-      <id>marceloverdijk</id>
-      <email>maceloverdijk@gmail.com</email>
-      <name>Marcel Overdijk</name>
-      <timezone>CET</timezone>
-    </developer>
-    <developer>
-      <id>rui.figueira</id>
-      <email>rui.figueira@gmail.com</email>
-      <name>Rui Figueira</name>
-      <timezone>WET</timezone>
-    </developer>
-  </developers>
-
-  <dependencies>
-
-    <dependency>
-      <groupId>org.togglz</groupId>
-      <artifactId>togglz-spring-boot-common-autoconfigure</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.togglz</groupId>
-      <artifactId>togglz-console</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-actuator</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-   <dependency>
-      <groupId>org.springframework</groupId>
-      <artifactId>spring-webmvc</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>org.jboss.spec.javax.servlet</groupId>
-      <artifactId>jboss-servlet-api_3.0_spec</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-configuration-processor</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <artifactId>spring-boot-starter-test</artifactId>
-      <groupId>org.springframework.boot</groupId>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <artifactId>spring-boot-starter-logging</artifactId>
-      <groupId>org.springframework.boot</groupId>
-      <scope>test</scope>
-    </dependency>
-
-  </dependencies>
-</project>
