--- spring-boot/boot2-autoconfigure/pom.xml
@@ -1,97 +1,0 @@
-<?xml version="1.0"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.togglz</groupId>
-    <artifactId>togglz-spring-boot</artifactId>
-    <version>2.6.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>togglz-spring-boot2-autoconfigure</artifactId>
-  <name>Togglz - Spring Boot 2 Autoconfiguration</name>
-  <description>Togglz - Spring Boot 2 Autoconfiguration</description>
-
-  <properties>
-    <spring-boot.version>2.0.1.RELEASE</spring-boot.version>
-
-    <!-- Spring Boot 2 only supports Java 8 and Spring 5 -->
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
-    <spring.version>5.0.5.RELEASE</spring.version>
-    <assertj.version>3.9.1</assertj.version>
-  </properties>
-
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <!-- Import dependency management from Spring Boot -->
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-dependencies</artifactId>
-        <version>${spring-boot.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-      <dependency>
-          <groupId>org.assertj</groupId>
-          <artifactId>assertj-core</artifactId>
-          <version>${assertj.version}</version>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-
-  <developers>
-    <developer>
-      <id>rui.figueira</id>
-      <email>rui.figueira@gmail.com</email>
-      <name>Rui Figueira</name>
-      <timezone>WET</timezone>
-    </developer>
-  </developers>
-
-  <dependencies>
-
-    <dependency>
-      <groupId>org.togglz</groupId>
-      <artifactId>togglz-spring-boot-common-autoconfigure</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.togglz</groupId>
-      <artifactId>togglz-console</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-actuator-autoconfigure</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>org.jboss.spec.javax.servlet</groupId>
-      <artifactId>jboss-servlet-api_3.0_spec</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <artifactId>spring-boot-starter-test</artifactId>
-      <groupId>org.springframework.boot</groupId>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <artifactId>spring-boot-starter-web</artifactId>
-      <groupId>org.springframework.boot</groupId>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <artifactId>spring-boot-starter-logging</artifactId>
-      <groupId>org.springframework.boot</groupId>
-      <scope>test</scope>
-    </dependency>
-
-  </dependencies>
-</project>
