+++ spring-boot/legacy-starter/pom.xml
@@ -9,18 +9,12 @@
     <version>2.6.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>togglz-spring-boot-starter</artifactId>
-  <name>Togglz - Spring Boot 2 Starter</name>
-  <description>Togglz - Spring Boot 2 Starter</description>
+  <artifactId>togglz-legacy-spring-boot-starter</artifactId>
+  <name>Togglz - Spring Boot 1.5 Autoconfigure</name>
+  <description>Togglz - Spring Boot 1.5 Autoconfigure</description>
 
   <properties>
-    <spring-boot.version>2.0.1.RELEASE</spring-boot.version>
-
-    <!-- Spring Boot 2 only supports Java 8 and Spring 5 -->
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
-    <spring.version>5.0.5.RELEASE</spring.version>
-    <assertj.version>3.9.1</assertj.version>
+    <spring-boot.version>1.5.12.RELEASE</spring-boot.version>
   </properties>
 
   <dependencyManagement>
@@ -33,15 +27,16 @@
         <type>pom</type>
         <scope>import</scope>
       </dependency>
-      <dependency>
-          <groupId>org.assertj</groupId>
-          <artifactId>assertj-core</artifactId>
-          <version>${assertj.version}</version>
-      </dependency>
     </dependencies>
   </dependencyManagement>
 
   <developers>
+    <developer>
+      <id>marceloverdijk</id>
+      <email>maceloverdijk@gmail.com</email>
+      <name>Marcel Overdijk</name>
+      <timezone>CET</timezone>
+    </developer>
     <developer>
       <id>rui.figueira</id>
       <email>rui.figueira@gmail.com</email>
@@ -65,7 +60,13 @@
 
     <dependency>
       <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-actuator-autoconfigure</artifactId>
+      <artifactId>spring-boot-actuator</artifactId>
+      <optional>true</optional>
+    </dependency>
+
+   <dependency>
+      <groupId>org.springframework</groupId>
+      <artifactId>spring-webmvc</artifactId>
       <optional>true</optional>
     </dependency>
 
@@ -76,13 +77,13 @@
     </dependency>
 
     <dependency>
-      <artifactId>spring-boot-starter-test</artifactId>
       <groupId>org.springframework.boot</groupId>
-      <scope>test</scope>
+      <artifactId>spring-boot-configuration-processor</artifactId>
+      <optional>true</optional>
     </dependency>
 
     <dependency>
-      <artifactId>spring-boot-starter-web</artifactId>
+      <artifactId>spring-boot-starter-test</artifactId>
       <groupId>org.springframework.boot</groupId>
       <scope>test</scope>
     </dependency>
