+++ spring-boot/legacy-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,6 @@
+# Auto Configure
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.togglz.legacy.spring.boot.autoconfigure.actuator.TogglzEndpointAutoConfiguration
+
+org.springframework.boot.actuate.autoconfigure.ManagementContextConfiguration=\
+org.togglz.legacy.spring.boot.autoconfigure.management.TogglzManagementContextConfiguration
