+++ spring-boot/legacy-starter/template.mf
