+++ spring-boot/starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,6 @@
+# Auto Configure
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.togglz.spring.boot.autoconfigure.actuator.TogglzEndpointAutoConfiguration
+
+org.springframework.boot.actuate.autoconfigure.web.ManagementContextConfiguration=\
+org.togglz.spring.boot.autoconfigure.management.TogglzManagementContextConfiguration
