+++ spring-boot/starter/template.mf
@@ -7,4 +7,10 @@ Bundle-Name: ${project.name}
 Version-Patterns:
  default;pattern="[=.=.=, =.+1.0)"
 Import-Template:
+ com.github.heneke.thymeleaf.togglz.*;version="0",
+ javax.servlet.*;version="0",
+ javax.validation.*;version="0",
+ org.springframework.*;version="[5.0,6.0)",
+ org.springframework.boot.*;version="[2.0,3.0)",
+ org.springframework.security.*;version="[5.0,6.0)",
  org.togglz.*;version="${osgi.bundles.version:default}"
